Moved Annotations to marlowe-object and fix merge problems
