+++ examples/nodejs/src/experimental-features/annotations.ts
@@ -1,9 +1,11 @@
-import * as t from "io-ts/lib/index.js";
-import * as R from "fp-ts/lib/Record.js";
-import * as O from "fp-ts/lib/Option.js";
 import * as M from "fp-ts/lib/Map.js";
 
-import { ContractBundleMap, bundleMapToList } from "@marlowe.io/marlowe-object";
+import {
+  ContractBundleMap,
+  bundleMapToList,
+  isAnnotated,
+  stripAnnotations,
+} from "@marlowe.io/marlowe-object";
 import {
   CreateContractRequestBase,
   RuntimeLifecycle,
@@ -23,37 +25,6 @@ import {
 import { RestClient } from "@marlowe.io/runtime-rest-client";
 import { ContractId, TxId } from "@marlowe.io/runtime-core";
 import { deepEqual } from "@marlowe.io/adapter/deep-equal";
-
-export interface Annotated<T> {
-  annotation: T;
-}
-
-export function isAnnotated(value: unknown): value is Annotated<unknown> {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    "annotation" in value &&
-    typeof value["annotation"] !== "undefined"
-  );
-}
-
-export const AnnotatedGuard = <T>(guard: t.Type<T>): t.Type<Annotated<T>> =>
-  t.type({ annotation: guard });
-
-export function stripAnnotations<T>(value: T): T {
-  if (typeof value !== "object" || value === null) {
-    return value;
-  }
-  if (value instanceof String) {
-    return `${value}` as T;
-  }
-  if (Array.isArray(value)) {
-    return value.map(stripAnnotations) as T;
-  }
-  return R.filterMapWithIndex((key, v) =>
-    key === "annotation" ? O.none : O.some(stripAnnotations(v))
-  )(value as any) as T;
-}
 
 function annotateInputFromClosure(contractClosure: ContractClosure) {
   return function (input: Core.Input): Core.Input {
@@ -93,11 +64,9 @@ async function annotatedClosure<T>(
   sourceObjectMap: ContractBundleMap<T>
 ): Promise<ContractClosure> {
   const { contractSourceId, intermediateIds } =
-    await restClient.createContractSources(
-      stripAnnotations(
-        bundleMapToList(sourceObjectMap)
-      ) as Obj.ContractBundleList<undefined>
-    );
+    await restClient.createContractSources({
+      bundle: bundleMapToList(sourceObjectMap),
+    });
 
   const closure = await getContractClosure({ restClient })(contractSourceId);
 
@@ -242,8 +211,9 @@ export async function mkSourceMap<T>(
       });
     },
     contractInstanceOf: async (contractId: ContractId) => {
-      const contractDetails =
-        await lifecycle.restClient.getContractById(contractId);
+      const contractDetails = await lifecycle.restClient.getContractById({
+        contractId,
+      });
 
       const initialContract = await lifecycle.restClient.getContractSourceById({
         contractSourceId: closure.main,
